fix(KEP): driveCmd touchups

+++ src/Commands/Guild/Impl/KEP_driveCmdImpl.ts
@@ -1,10 +1,10 @@
 import { ApplicationCommandData, Collection, CommandInteraction, Message, Snowflake } from "discord.js";
-import moment from "moment";
+import moment from "moment-timezone";
 import { guildMap } from "../../..";
 import { drive as driveLink } from "../../../../values/KEP/info.json";
 import { commandLiteral } from "../../../Entities/Generic/command";
 import { fetchCommandID } from "../../../Queries/Generic/Commands";
-import { dropDrivePermission, findDrivePerm, saveDrivePermission } from "../../../Queries/KEP/drivePerms";
+import { dropDrivePermission, fetchDrivePermissions, findDrivePerm, saveDrivePermission } from "../../../Queries/KEP/Drive";
 import { fetchStudent } from "../../../Queries/KEP/Student";
 import { addDrivePermission, deleteDrivePermission } from "../../../tools/Google/Gdrive";
 import { scheduleTask } from "../../../tools/scheduler";
@@ -43,9 +43,9 @@ export class KEP_driveCmdImpl extends AbstractGuildCommand implements KEP_driveC
                 options: [
                     {
                         name: "duration",
-                        description: `Διάρκεια Πρόσβασης σε ώρες (default: ${defaultHours})`,
+                        description: `Διάρκεια Πρόσβασης σε ώρες (default: ${defaultHours} ώρες)`,
                         type: "INTEGER",
-                        choices: [3, 6, 12].map(i => ({ name: `${i}`, value: i })),
+                        choices: [3, 6, 12].map(i => ({ name: `${i} ώρες`, value: i })),
                         required: false
                     }
                 ]
@@ -56,21 +56,24 @@ export class KEP_driveCmdImpl extends AbstractGuildCommand implements KEP_driveC
     async interactiveExecute(interaction: CommandInteraction): Promise<unknown> {
         await interaction.deferReply({ ephemeral: true });
         const subCommand = interaction.options.getSubcommand();
-        const [durationHours] = enabledHours.map(i => interaction.options.getInteger(`${i}`)).filter(v => !!v);
-        console.log(`time is ${durationHours} hours`);
+        const inputHours = interaction.options.getInteger("duration");
         switch (subCommand) {
             case registerLiteral: {
                 const existingPermission = await findDrivePerm(interaction.user.id)
                 if (!!existingPermission)
-                    return interaction.editReply(`Έχετε ήδη ενεργοποιημένη πρόσβαση μέχρι ${moment(existingPermission.destroyedAt).format("hh:mm:ss a")}`)
-                const duration = durationHours ?? defaultHours;
+                    return interaction.editReply(`Έχετε ήδη ενεργοποιημένη πρόσβαση μέχρι **${moment(existingPermission.destroyedAt).tz("Europe/Athens").format("DD/MM dddd, kk:mm")}**`)
+                const duration = inputHours ?? defaultHours;
                 const student = await fetchStudent({ member_id: interaction.user.id });
+                if (!student)
+                    return interaction.editReply('Δεν είστε εγγεγραμμένος. Εγγραφείτε με /registration register');
+                const time = moment().tz("Europe/Athens").add(duration, "hours");
                 const perm = await addDrivePermission(student.email);
-                const time = moment().add(duration, "hours");
                 await saveDrivePermission(perm.data.id, time, interaction.user.id);
-                scheduleTask(time, () =>
-                    deleteDrivePermission(perm.data.id)
-                        .then(() => dropDrivePermission(perm.data.id))
+                scheduleTask(time, async () => {
+                    if (!!await fetchDrivePermissions(interaction.user.id)[0]) //check if record still exists
+                        deleteDrivePermission(perm.data.id)
+                            .then(() => dropDrivePermission(perm.data.id))
+                }
                 );
                 return interaction.editReply(`Ενεργοποιήθηκε η πρόσβαση σας για **${duration} ώρες**\nΟ σύνδεσμος: ${driveLink}`)
             }
